requirements.txt

airflow-dbt==0.4.0
apache-airflow-providers-postgres==6.1.3
dbt-common==1.12.0
dbt-core==1.8.2
dbt-postgres==1.8.2
google-api-core==2.24.0
google-auth==2.37.0
great-expectations==1.4.2
gspread==6.2.0
numpy==1.26.4
oauth2client==4.1.3
pandas==2.1.4
pipdeptree==2.26.1
psycopg2-binary==2.9.10
dlt==1.10.0
python-dotenv
requests
astronomer-cosmos[dbt.postgres]==1.10.0
